feat(chat): move chat feature from CE to EE
Chat is an enterprise-only feature but its code was located in the CE
code path. Move the backend module from src/app/chat/ to src/app/ee/chat/
and shared types from lib/automation/chat/ to lib/ee/chat/. Register the
chat module only for CLOUD and ENTERPRISE editions in app.ts.
