+++ packages/server/api/src/app/ee/chat/chat-approval-gate.ts
@@ -1,5 +1,5 @@
-import { redisConnections } from '../database/redis-connections'
-import { pubsub } from '../helper/pubsub'
+import { redisConnections } from '../../database/redis-connections'
+import { pubsub } from '../../helper/pubsub'
 
 const GATE_TIMEOUT_MS = 5 * 60 * 1000
 const CHANNEL_PREFIX = 'tool-approval:'
