+++ packages/server/api/src/app/ee/chat/chat-controller.ts
@@ -10,7 +10,7 @@ import { pipeUIMessageStreamToResponse } from 'ai'
 import { FastifyPluginAsyncZod } from 'fastify-type-provider-zod'
 import { StatusCodes } from 'http-status-codes'
 import { z } from 'zod'
-import { securityAccess } from '../core/security/authorization/fastify-security'
+import { securityAccess } from '../../core/security/authorization/fastify-security'
 import { chatApprovalGate } from './chat-approval-gate'
 import { chatService } from './chat-service'
