+++ packages/server/api/src/app/ee/chat/chat-conversation-entity.ts
@@ -1,6 +1,6 @@
 import { ChatConversation, Platform, Project } from '@activepieces/shared'
 import { EntitySchema } from 'typeorm'
-import { ApIdSchema, BaseColumnSchemaPart } from '../database/database-common'
+import { ApIdSchema, BaseColumnSchemaPart } from '../../database/database-common'
 
 type ChatConversationWithRelations = ChatConversation & {
     platform: Platform
