+++ packages/server/api/src/app/ee/chat/chat-file-utils.ts
