+++ packages/server/api/src/app/ee/chat/chat-model-factory.ts
