+++ packages/server/api/src/app/ee/chat/chat-service.ts
@@ -17,16 +17,16 @@ import {
 } from '@activepieces/shared'
 import { createUIMessageStream, LanguageModel, ModelMessage, stepCountIs, streamText } from 'ai'
 import { FastifyBaseLogger } from 'fastify'
-import { aiProviderService } from '../ai/ai-provider-service'
-import { repoFactory } from '../core/db/repo-factory'
-import { buildPaginator } from '../helper/pagination/build-paginator'
-import { paginationHelper } from '../helper/pagination/pagination-utils'
-import { Order } from '../helper/pagination/paginator'
-import { system } from '../helper/system/system'
-import { AppSystemProp } from '../helper/system/system-props'
-import { mcpProjectSelection } from '../mcp/mcp-project-selection'
-import { projectService } from '../project/project-service'
-import { userService } from '../user/user-service'
+import { aiProviderService } from '../../ai/ai-provider-service'
+import { repoFactory } from '../../core/db/repo-factory'
+import { buildPaginator } from '../../helper/pagination/build-paginator'
+import { paginationHelper } from '../../helper/pagination/pagination-utils'
+import { Order } from '../../helper/pagination/paginator'
+import { system } from '../../helper/system/system'
+import { AppSystemProp } from '../../helper/system/system-props'
+import { mcpProjectSelection } from '../../mcp/mcp-project-selection'
+import { projectService } from '../../project/project-service'
+import { userService } from '../../user/user-service'
 import { chatCompaction } from './chat-compaction'
 import { ChatConversationEntity } from './chat-conversation-entity'
 import { buildUserContentWithFiles } from './chat-file-utils'
