+++ packages/server/api/src/app/ee/chat/prompt/chat-prompt.ts
