+++ packages/server/api/src/app/ee/chat/tools/chat-tools.ts
@@ -2,13 +2,13 @@ import { FlowRunStatus, FlowStatus, Project, RunEnvironment } from '@activepiece
 import { tool } from 'ai'
 import { FastifyBaseLogger } from 'fastify'
 import { z } from 'zod'
-import { appConnectionService } from '../../app-connection/app-connection-service/app-connection-service'
-import { flowService } from '../../flows/flow/flow.service'
-import { flowRunService } from '../../flows/flow-run/flow-run-service'
-import { formatFlowLine } from '../../mcp/tools/ap-list-flows'
-import { executeAdhocAction, formatRunSummary } from '../../mcp/tools/flow-run-utils'
-import { mcpUtils } from '../../mcp/tools/mcp-utils'
-import { tableService } from '../../tables/table/table.service'
+import { appConnectionService } from '../../../app-connection/app-connection-service/app-connection-service'
+import { flowService } from '../../../flows/flow/flow.service'
+import { flowRunService } from '../../../flows/flow-run/flow-run-service'
+import { formatFlowLine } from '../../../mcp/tools/ap-list-flows'
+import { executeAdhocAction, formatRunSummary } from '../../../mcp/tools/flow-run-utils'
+import { mcpUtils } from '../../../mcp/tools/mcp-utils'
+import { tableService } from '../../../tables/table/table.service'
 import { chatPrompt } from '../prompt/chat-prompt'
 
 const RESOURCE_TYPES = ['flows', 'tables', 'runs', 'connections'] as const
