+++ packages/shared/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@activepieces/shared",
-  "version": "0.71.5",
+  "version": "0.71.6",
   "type": "commonjs",
   "sideEffects": false,
   "main": "./dist/src/index.js",
