+++ packages/shared/src/index.ts
@@ -82,7 +82,7 @@ export * from './lib/automation/forms'
 export * from './lib/automation/mcp'
 export * from './lib/automation/agents'
 export * from './lib/automation/knowledge-base'
-export * from './lib/automation/chat'
+export * from './lib/ee/chat'
 export * from './lib/automation/tables'
 export * from './lib/automation/project-release/project-release'
 export * from './lib/automation/project-release/project-release.request'
