+++ packages/shared/src/lib/ee/chat/index.ts
